+++ aspnetcore/src/ElasticSearchIndexer/Indexer.cs
@@ -1,4 +1,4 @@
-﻿using Api.DataAccess;
+﻿using Api.DataAccess.Repositories;
 using Api.Maps;
 using Api.Models.Entities;
 using Api.Models.FundingCall;
