+++ aspnetcore/src/ElasticSearchIndexer/Program.cs
@@ -1,5 +1,5 @@
 ﻿using Api.ConfigurationExtensions;
-using Api.DataAccess;
+using Api.DataAccess.Repositories;
 using Api.DatabaseContext;
 using Api.Maps;
 using Api.Models.Entities;
